expose preprocess prop
Currently expects a function that returns an object with `type`, `name` and `source` keys. As such, this isn’t compatible with the `svelte.preprocess` syntax.

Would be better to hook into `svelte.preprocess` for greater flexibility but that happens in the compiler webworker so we cannot pass functions directly. Work-around would be blob + object url but that seems hacky. Mmm..